LINEトークをマスタから取得するように変更

+++ admin/register.php
@@ -3,33 +3,43 @@ require_once('../model/entity/GameInfo.php');
 require_once('../model/dao/GameInfoDao.php');
 use entity\GameInfo;
 use dao\GameInfoDao;
-$msg = '';
-if (isset($_POST['register'])) {
-    // 登録・修正''
-    $msg = '登録';
-    $gameInfo = new GameInfo(
-        $_POST['title']
-        , $_POST['game_date']
-        , $_POST['start_time']
-        , $_POST['end_time']
-        , $_POST['place']
-        , $_POST['limit_number']
-        , $_POST['detail']
-    );
-    
-    $gameInfoDao = new GameInfoDao();
-    if($_POST['id'] == '') {
-        $gameInfoDao->insert($gameInfo);
+
+session_start();
+
+if (isset($_POST["csrf_token"]) 
+ && $_POST["csrf_token"] === $_SESSION['csrf_token']) {
+
+    $msg = '';
+    if (isset($_POST['register'])) {
+        // 登録・修正''
+        $msg = '登録';
+        $gameInfo = new GameInfo(
+            $_POST['title']
+            , $_POST['game_date']
+            , $_POST['start_time']
+            , $_POST['end_time']
+            , $_POST['place']
+            , $_POST['limit_number']
+            , $_POST['detail']
+        );
+        
+        $gameInfoDao = new GameInfoDao();
+        if($_POST['id'] == '') {
+            $gameInfoDao->insert($gameInfo);
+        } else {
+            $gameInfo->id = $_POST['id'];
+            $gameInfoDao->update($gameInfo);
+        }
     } else {
-        $gameInfo->id = $_POST['id'];
-        $gameInfoDao->update($gameInfo);
+        $gameInfoDao = new GameInfoDao();
+        if($_POST['id'] != '') {
+            $msg = '削除';
+            $gameInfoDao->delete($_POST['id']);
+        }
     }
+    unset($_SESSION['csrf_token']);
 } else {
-    $gameInfoDao = new GameInfoDao();
-    if($_POST['id'] != '') {
-        $msg = '削除';
-        $gameInfoDao->delete($_POST['id']);
-    }
+    header('Location: ./index.php');
 }
 ?>
